+++ pyam/statistics.py
@@ -4,6 +4,7 @@ import numpy as np
 import pandas as pd
 
 from pyam import filter_by_meta
+from pyam.exceptions import deprecation_warning
 from pyam.str import is_str
 from pyam.utils import META_IDX, is_list_like
 
@@ -30,7 +31,12 @@ class Statistics:
     """
 
     def __init__(  # noqa: C901
-        self, df, groupby=None, filters=None, rows=False, percentiles=[0.25, 0.5, 0.75]
+        self,
+        df,
+        groupby=None,
+        filters=None,
+        rows=False,
+        percentiles=[0.05, 0.25, 0.5, 0.75, 0.95],
     ):
         self.df = df
         self.idx_depth = None
@@ -231,71 +237,111 @@ class Statistics:
             return ret
 
     def summarize(
-        self, center="mean", fullrange=None, interquartile=None, custom_format="{:.2f}"
+        self,
+        center="mean",
+        *,
+        limits=None,
+        fullrange=None,
+        interquartile=None,
+        custom_format="{:.2f}",
     ):
         """Format the compiled statistics to a concise string output
 
         Parameters
         ----------
-        center : str, default `mean`
-            what to return as 'center' of the summary: `mean`, `50%`, `median`
-        fullrange : bool, default None
-            return full range of data if True or `fullrange`, `interquartile`
-            and `format_spec` are None
-        interquartile : bool, default None
-            return interquartile range if True
+        center : str, optional
+            The 'center' of the summary: 'mean', 'median' (50th percentile).
+        limits: str, optional
+            The upper and lower limits of the shown range: 'full range',
+            'interquartile', 'central 90%', defaults to 'full range'.
+        fullrange : bool, optional, deprecated
+            Return full range of data if True or `fullrange`, `interquartile`
+            and `format_spec` are None.
+        interquartile : bool, optional, deprecated
+            Return interquartile range if True.
         custom_format : formatting specifications
         """
-        # call `reindex()` to reorder index and columns
+        # call `reindex()` to reorder index and columns before summarizing
         self.reindex(copy=False)
 
-        center = "median" if center == "50%" else center
-        if fullrange is None and interquartile is None:
-            fullrange = True
-        return self.stats.apply(
+        # TODO: deprecated, remove "50%" for release >= 4.1 to streamline signature
+        if center == "50%":
+            deprecation_warning(
+                "use `center='median'` instead.",
+                "Using `center='50%'",
+            )
+            center = "median"
+
+        # TODO: remove deprecated args for release >= 4.1, remove all code below
+        if fullrange is not None and interquartile is not None:
+            raise ValueError(
+                "Cannot use `interquartile` and `fullrange`, use `limits` instead."
+            )
+
+        for arg, name, value in [
+            (fullrange, "fullrange", "full range"),
+            (interquartile, "interquartile", "interquartile"),
+        ]:
+            if arg is not None:
+                if limits is not None:
+                    raise ValueError(
+                        f"Cannot use `{name}` with `limits`, "
+                        f"use `limits='{value}'` instead."
+                    )
+                deprecation_warning(
+                    f"use `limits='{value}'` instead.",
+                    f"The argument `{name}`",
+                )
+                limits = value
+
+        if limits is None:
+            limits = "full range"
+        # TODO remove all code up to here, set `limits='full range'` as default
+        if limits not in ["full range", "interquartile", "central 90%"]:
+            raise ValueError(f"Invalid argument for `limits`: {limits}")
+
+        summary = self.stats.apply(
             format_rows,
             center=center,
-            fullrange=fullrange,
-            interquartile=interquartile,
+            limits=limits,
             custom_format=custom_format,
             axis=1,
             raw=False,
         )
+        summary[("count", "")] = summary[("count", "")].map(int)
+        return summary
 
 
-# %% auxiliary functions
-
-
-def format_rows(
-    row, center, fullrange=None, interquartile=None, custom_format="{:.2f}"
-):
+def format_rows(row, center, limits, custom_format="{:.2f}"):
     """Format a row with `describe()` columns to a concise string"""
-    if (fullrange or 0) + (interquartile or 0) == 1:
-        legend = "{} ({})".format(
-            center, "max, min" if fullrange is True else "interquartile range"
-        )
 
-        row_index = row.index.droplevel(2).drop_duplicates()
-        ret_index = pd.MultiIndex.from_tuples([("count", "")]).append(row_index)
-        ret_index.names = [None, legend]
-    else:
-        raise ValueError("Use either fullrange or interquartile range.")
+    legend = center
+    if limits == "full range":
+        legend += " (max, min)"
+        upper, lower = ("max", "min")
+    elif limits == "interquartile":
+        legend += " (p75, p25)"
+        upper, lower = ("75%", "25%")
+    elif limits == "central 90%":
+        legend += " (p95, p5)"
+        upper, lower = ("95%", "5%")
 
-    ret = pd.Series(index=ret_index, dtype=float)
+    row_index = row.index.droplevel(2).drop_duplicates()
+    ret_index = pd.MultiIndex.from_tuples([("count", "")]).append(row_index)
+    ret_index.names = [None, legend]
+
+    ret = pd.Series(index=ret_index, dtype=object)
 
     row = row.sort_index()
-    center = "50%" if center == "median" else center
 
     # get maximum of `count` and write to first entry of return series
     count = max(
         [i for i in row.loc[(slice(None), slice(None), "count")] if not np.isnan(i)]
     )
-    ret.loc[("count", "")] = (f"{count:.0f}") if count > 1 else ""
-
-    # set upper and lower for the range
-    upper, lower = ("max", "min") if fullrange is True else ("75%", "25%")
+    ret.loc[("count", "")] = count
 
     # format `describe()` columns to string output
+    center = "50%" if center == "median" else center
     for i in row_index:
         x = row.loc[i]
         _count = x["count"]
@@ -306,7 +352,7 @@ def format_rows(
                 x[center], x[upper], x[lower]
             )
         elif _count == 1:
-            s = f"{custom_format}".format(x["50%"])
+            s = f"{custom_format}".format(x[center])
         # add count of this section as `[]` if different from count_max
         if 0 < _count < count:
             s += f" [{_count:.0f}]"