+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Next release
 
+## API changes
+
+The method `Statistics.summarize()` includes a new argument `limits` that can take the
+values 'full range', 'interquartile' or 'central 90%'. The previous boolean arguments
+`fullrange` and `interquartile` are marked as deprecated.
+
+## Individual updates
+
+- [#991](https://github.com/IAMconsortium/pyam/pull/991) Support the central 90% as limits in the statistics summary
 - [#989](https://github.com/IAMconsortium/pyam/pull/989) Add an `IamDataFrame.series` attribute to get timeseries data
   as **pd.Series**